Final commit
What a weird project. Had a lot of problems with webpack's localhost,
and then ran into a huge issue with git that deleted my commit history,
luckily I only had to rewrite the easiest of the files to rewrite. I
will be more cognizant to push all my commits in the future, as opposed
to keeping them on the device and pushing them all at once. It would
have been really bad if my binarySearchTree file were to have been
deleted in that process. Overall, really fun project, and the
visual of the tree that prettyPrint provided was a really nice touch.
On to the next!